% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anomalies_data.R
\name{get_anomalies_data}
\alias{get_anomalies_data}
\title{Function to get anomalies data from s3}
\usage{
get_anomalies_data()
}
\value{
tibble dataframe
}
\description{
Function to get anomalies data from s3
}
