% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anomalies_summary_by_date.R
\name{get_anomalies_summary_by_date}
\alias{get_anomalies_summary_by_date}
\title{Function to get household data from s3}
\usage{
get_anomalies_summary_by_date()
}
\value{
tibble dataframe
}
\description{
Function to get household data from s3
}
