% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object_update_time.R
\name{get_object_update_time}
\alias{get_object_update_time}
\title{Function to get household data from s3}
\usage{
get_object_update_time()
}
\value{
tibble dataframe
}
\description{
Function to get household data from s3
}
