% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registration_forms.R
\name{get_registration_forms}
\alias{get_registration_forms}
\title{Function to get household data from s3}
\usage{
get_registration_forms()
}
\value{
tibble dataframe
}
\description{
Function to get household data from s3
}
