% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s3_data.R
\name{get_s3_data}
\alias{get_s3_data}
\title{Utility function to retrieve Kwale filepath}
\usage{
get_s3_data(s3obj = NULL, bucket, object_key, filename)
}
\description{
Utility function to retrieve Kwale filepath
}
