% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress.R
\name{mod_progress_ui}
\alias{mod_progress_ui}
\title{mod_map.R}
\usage{
mod_progress_ui(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
